freezing folder: fix describe description, rename file
