+++ test/unit/operations/write/freezing/freezesComplexValuesReferencedFromPayload.ts
@@ -1,43 +1,33 @@
 import { CacheContext } from '../../../../../src/context';
 import { GraphSnapshot } from '../../../../../src/GraphSnapshot';
 import { write } from '../../../../../src/operations/write';
-import { StaticNodeId } from '../../../../../src/schema';
 import { query, strictConfig } from '../../../../helpers';
-
-const { QueryRoot: QueryRootId } = StaticNodeId;
 
 describe(`operations.write`, () => {
 
   const context = new CacheContext(strictConfig);
   const empty = new GraphSnapshot();
-  const nestedQuery = query(`{
-    foo {
-      bar {
-        baz
-      }
-    }
-  }`);
 
-  describe(`snapshot freezing`, () => {
+  describe(`freezes complex values referenced from the payload`, () => {
 
-    it(`freezes snapshots after writing`, () => {
-      const snapshot = write(context, empty, nestedQuery, {
+    it(`checks values references from the payload`, () => {
+      const simpleQuery = query(`{ foo }`);
+      const payload = {
         foo: {
           bar: [
             { baz: 123 },
             { baz: 321 },
           ],
         },
-      }).snapshot;
+      };
+      write(context, empty, simpleQuery, payload);
 
       expect(() => {
-        const root = snapshot.getNodeData(QueryRootId);
-        root.foo.bar[0].baz = 111;
+        payload.foo.bar[0].baz = 111;
       }).to.throw(/property.*baz/);
 
       expect(() => {
-        const root = snapshot.getNodeData(QueryRootId);
-        root.foo.fizz = 'nope';
+        (payload.foo as any).fizz = 'nope';
       }).to.throw(/property.*fizz/);
     });
